refactoring and code cleanup
