+++ src/main/java/com/cribhub/backend/controllers/CribController.java
@@ -2,7 +2,7 @@ package com.cribhub.backend.controllers;
 
 import com.cribhub.backend.domain.Crib;
 import com.cribhub.backend.dto.CribDTO;
-import com.cribhub.backend.services.CribService;
+import com.cribhub.backend.services.intefaces.CribService;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
