+++ src/main/java/com/cribhub/backend/controllers/CustomerController.java
@@ -7,9 +7,9 @@ import com.cribhub.backend.domain.Crib;
 import com.cribhub.backend.domain.Customer;
 import com.cribhub.backend.dto.CustomerDTO;
 import com.cribhub.backend.dto.CustomerUpdateDTO;
-import com.cribhub.backend.services.CribService;
-import com.cribhub.backend.services.CustomerService;
 import com.cribhub.backend.services.CustomerServiceImpl;
+import com.cribhub.backend.services.intefaces.CribService;
+import com.cribhub.backend.services.intefaces.CustomerService;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Min;
 import lombok.extern.log4j.Log4j2;
