+++ src/test/java/com/cribhub/backend/controller/CustomerControllerTests.java
@@ -8,8 +8,8 @@ import com.cribhub.backend.controllers.exceptions.UsernameAlreadyTakenException;
 import com.cribhub.backend.domain.Crib;
 import com.cribhub.backend.domain.Customer;
 import com.cribhub.backend.dto.CustomerDTO;
-import com.cribhub.backend.services.CribService;
 import com.cribhub.backend.services.CustomerServiceImpl;
+import com.cribhub.backend.services.intefaces.CribService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
