+++ src/test/java/com/cribhub/backend/integration/CustomerControllerIntegrationTest.java
@@ -6,8 +6,8 @@ import com.cribhub.backend.controllers.exceptions.UsernameAlreadyTakenException;
 import com.cribhub.backend.domain.Crib;
 import com.cribhub.backend.domain.Customer;
 import com.cribhub.backend.dto.CustomerDTO;
-import com.cribhub.backend.services.CribService;
-import com.cribhub.backend.services.CustomerService;
+import com.cribhub.backend.services.intefaces.CribService;
+import com.cribhub.backend.services.intefaces.CustomerService;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
